refactor(security): extract scan detectors under line budget
Move MCP, harness-wiring, package.json, GitHub Actions, and Python
project document scanners into scan_detectors.py so scan_engine.py
stays under the phase257 2000-line production ceiling with headroom
for follow-on coalescing work. Facade re-exports and monkeypatch sync
are unchanged.
