+++ tests/test_phase257_facades.py
@@ -37,6 +37,7 @@ FAMILIES = {
             "enrichment",
             "suppression",
             "template_audit",
+            "scan_detectors",
             "scan_engine",
             "commands",
         ],
